+++ changelog.md
@@ -2,6 +2,7 @@
 #### Fixed
 - Adding new patient with external id in a database which contains a patient with the same external id 
   and empty fields causes a conflict, if the NullMatcher has been activated
+- Set tomcat connector's scheme depending on ML_REVERSEPROXY_SSL environment variable
 
 ### 1.9.0 - 2020-12-21
 #### Added